feat(web): merge branches, consensus namings, refactors, scrolltop
