+++ web/src/components/RegistriesDisplay/RegistriesGrid.tsx
@@ -14,7 +14,7 @@ import RegistryCard from "components/RegistryCard";
 const GridContainer = styled.div`
   --gap: 24px;
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(min(100%, max(348px, (100% - var(--gap) * 2)/3)), 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(min(100%, max(274px, (100% - var(--gap) * 2)/3)), 1fr));
   align-items: center;
   gap: var(--gap);
 `;
