+++ web/src/components/RegistryCard/RegistryInfo.tsx
