--- web/src/consts/statuses.ts
@@ -1,6 +1,0 @@
-export enum Statuses {
-  pending,
-  disputed,
-  included,
-  removed,
-}
